requirements.txt
Flask
Flask-SQLAlchemy
psycopg2
google-cloud-vision
python-dotenv
